+++ .cargo/config.toml
@@ -1,6 +1,6 @@
-# Just used for rust-analyzer.
+# # Just used for rust-analyzer.
 # [build]
 # target = "riscv64gc-unknown-none-elf"
-# target = 'aarch64-unknown-none-softfloat'
-# target = 'x86_64-unknown-none'
-# target = 'loongarch64-unknown-none'
+# # target = 'aarch64-unknown-none-softfloat'
+# # target = 'x86_64-unknown-none'
+# # target = 'loongarch64-unknown-none'
