+++ src/lib.rs
@@ -17,6 +17,11 @@ extern crate cfg_if;
 #[macro_use]
 mod utils;
 
+mod common;
+pub use common::addr::*;
+pub use common::debug::DebugConsole;
+pub use common::page::PageAlloc;
+
 cfg_if! {
     if #[cfg(feature = "libos")] {
         #[path = "libos/mod.rs"]
@@ -24,14 +29,11 @@ cfg_if! {
 
         pub use polyhal_macro::{arch_entry, arch_interrupt};
         pub use imp::context::*;
-        pub use imp::debug::DebugConsole;
-        pub use imp::page::PageAlloc;
         pub use imp::addr::*;
         pub use imp::init;
-        pub use imp::mem::free_pmem_regions;
+        pub use imp::mem::get_mem_areas;
         pub use imp::vm::{Page, PageTable, GenericPageTable};
         pub use imp::addr::MMUFlags;
-        pub use imp::vm::PageSize;
         pub use imp::mem::{pmem_read, pmem_copy, pmem_write, pmem_zero};
     } else {
         #[path = "bare/mod.rs"]
@@ -41,20 +43,14 @@ cfg_if! {
         use imp::consts::*;
         use imp::multicore::MultiCore;
         use imp::*;
-        use imp::current_arch::*;
-        use imp::debug::*;
-        use imp::time::*;
 
         pub use imp::{
-            addr::*,
-            page::PageAlloc,
             pagetable::{PageTable, PageTableWrapper, MappingFlags, MappingSize}
         };
         pub use imp::{
             get_mem_areas, init,
             TrapFrameArgs, TrapType, PAGE_SIZE,
             time::*,
-            debug::DebugConsole,
             current_arch::{
                 run_user_task, shutdown, kernel_page_table,
                 TrapFrame, VIRT_ADDR_START,
