+++ src/libos/addr.rs
@@ -1,17 +1,27 @@
 use bitflags::bitflags;
-use crate::bit;
-
-/// Physical address.
-pub type PhysAddr = usize;
-
-/// Virtual address.
-pub type VirtAddr = usize;
-
-/// Device address.
-pub type DevVAddr = usize;
+use crate::{bit, PhysAddr, VirtAddr};
+use crate::common::addr::PhysPage;
 
 pub const PAGE_SIZE: usize = 0x1000;
+pub const PAGE_BITS: usize = 12;
 
+impl PhysPage {
+    pub fn addr(&self) -> PhysAddr {
+        PhysAddr::new(self.0 << PAGE_BITS)
+    }
+}
+
+impl VirtAddr {
+    pub fn add_offset(&mut self, s: usize) {
+        self.0 += s;
+    }
+}
+
+impl PhysAddr {
+    pub fn add_offset(&mut self, s: usize) {
+        self.0 += s;
+    }
+}
 
 pub const fn align_down(addr: usize) -> usize {
     addr & !(PAGE_SIZE - 1)
@@ -33,14 +43,6 @@ pub const fn page_offset(addr: usize) -> usize {
     addr & (PAGE_SIZE - 1)
 }
 
-/// The error type which is returned from HAL functions.
-/// TODO: more error types.
-#[derive(Debug)]
-pub struct HalError;
-
-/// The result type returned by HAL functions.
-pub type HalResult<T = ()> = core::result::Result<T, HalError>;
-
 bitflags! {
     /// Generic memory flags.
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -50,6 +52,5 @@ bitflags! {
         const WRITE     = bit!(3);
         const EXECUTE   = bit!(4);
         const USER      = bit!(5);
-        const HUGE_PAGE = bit!(6);
     }
 }
