+++ src/libos/mod.rs
@@ -1,25 +1,24 @@
 pub mod api;
 pub mod context;
 pub mod debug;
-pub mod page;
 pub mod mem;
 pub mod vm;
 pub mod mock_mem;
 pub mod addr;
-use crate::utils::init_once::InitOnce;
-use page::PageAlloc;
+use crate::utils::once::LazyInit;
+use crate::PageAlloc;
 
 #[no_mangle]
 fn main() {
     unsafe { api::_main_for_arch(0) };
 }
 
-pub(crate) static PAGE_ALLOC: InitOnce<&dyn PageAlloc> = InitOnce::new();
+pub(crate) static PAGE_ALLOC: LazyInit<&dyn PageAlloc> = LazyInit::new();
 
 /// Init arch with page allocator, like log crate
 /// Please initialize the allocator before calling this function.
 pub fn init(page_alloc: &'static dyn PageAlloc) {
-    PAGE_ALLOC.init_once_by(page_alloc);
+    PAGE_ALLOC.init_by(page_alloc);
 
     self::mem::init_mock_mem();
 }
