--- src/libos/page.rs
@@ -1,6 +1,0 @@
-use crate::PhysAddr;
-
-pub trait PageAlloc: Sync {
-    fn alloc(&self) -> PhysAddr;
-    fn dealloc(&self, ppn: PhysAddr);
-}
