+++ src/libos/vm.rs
@@ -1,8 +1,5 @@
-//! Virtual memory operations.
-
 use super::mem::{MOCK_PHYS_MEM, PMEM_MAP_VADDR, PMEM_SIZE};
 use crate::{is_aligned, MMUFlags, PhysAddr, VirtAddr, PAGE_SIZE};
-
 
 /// Errors may occur during address translation.
 #[derive(Debug)]
@@ -31,44 +28,15 @@ impl<T> IgnoreNotMappedErr for PagingResult<T> {
     }
 }
 
-/// Possible page size (4K, 2M, 1G).
-#[repr(usize)]
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-pub enum PageSize {
-    Size4K = 0x1000,
-    Size2M = 0x20_0000,
-    Size1G = 0x4000_0000,
-}
-
 /// A 4K, 2M or 1G size page.
 #[derive(Debug, Copy, Clone)]
 pub struct Page {
     pub vaddr: VirtAddr,
-    pub size: PageSize,
-}
-
-impl PageSize {
-    pub const fn is_aligned(self, addr: usize) -> bool {
-        self.page_offset(addr) == 0
-    }
-
-    pub const fn align_down(self, addr: usize) -> usize {
-        addr & !(self as usize - 1)
-    }
-
-    pub const fn page_offset(self, addr: usize) -> usize {
-        addr & (self as usize - 1)
-    }
-
-    pub const fn is_huge(self) -> bool {
-        matches!(self, Self::Size1G | Self::Size2M)
-    }
 }
 
 impl Page {
-    pub fn new_aligned(vaddr: VirtAddr, size: PageSize) -> Self {
-        debug_assert!(size.is_aligned(vaddr));
-        Self { vaddr, size }
+    pub fn new(vaddr: VirtAddr) -> Self {
+        Self { vaddr }
     }
 }
 
@@ -103,18 +71,10 @@ pub trait GenericPageTable: Sync + Send {
     fn map(&mut self, page: Page, paddr: PhysAddr, flags: MMUFlags) -> PagingResult;
 
     /// Unmap the page of `vaddr`.
-    fn unmap(&mut self, vaddr: VirtAddr) -> PagingResult<(PhysAddr, PageSize)>;
-
-    /// Change the `flags` of the page of `vaddr`.
-    fn update(
-        &mut self,
-        vaddr: VirtAddr,
-        paddr: Option<PhysAddr>,
-        flags: Option<MMUFlags>,
-    ) -> PagingResult<PageSize>;
+    fn unmap(&mut self, vaddr: VirtAddr) -> PagingResult<PhysAddr>;
 
     /// Query the physical address which the page of `vaddr` maps to.
-    fn query(&self, vaddr: VirtAddr) -> PagingResult<(PhysAddr, MMUFlags, PageSize)>;
+    fn query(&self, vaddr: VirtAddr) -> PagingResult<(PhysAddr, MMUFlags)>;
 
     fn map_cont(
         &mut self,
@@ -123,87 +83,50 @@ pub trait GenericPageTable: Sync + Send {
         start_paddr: PhysAddr,
         flags: MMUFlags,
     ) -> PagingResult {
-        assert!(is_aligned(start_vaddr));
-        assert!(is_aligned(start_vaddr));
+        assert!(is_aligned(start_vaddr.addr()));
+        assert!(is_aligned(start_vaddr.addr()));
         assert!(is_aligned(size));
-        debug!(
-            "map_cont: {:#x?} => {:#x}, flags={:?}",
-            start_vaddr..start_vaddr + size,
-            start_paddr,
-            flags
-        );
         let mut vaddr = start_vaddr;
         let mut paddr = start_paddr;
-        let end_vaddr = vaddr + size;
-        if flags.contains(MMUFlags::HUGE_PAGE) {
-            while vaddr < end_vaddr {
-                let remains = end_vaddr - vaddr;
-                let page_size = if remains >= PageSize::Size1G as usize
-                    && PageSize::Size1G.is_aligned(vaddr)
-                    && PageSize::Size1G.is_aligned(paddr)
-                {
-                    PageSize::Size1G
-                } else if remains >= PageSize::Size2M as usize
-                    && PageSize::Size2M.is_aligned(vaddr)
-                    && PageSize::Size2M.is_aligned(paddr)
-                {
-                    PageSize::Size2M
-                } else {
-                    PageSize::Size4K
-                };
-                let page = Page::new_aligned(vaddr, page_size);
-                self.map(page, paddr, flags)?;
-                vaddr += page_size as usize;
-                paddr += page_size as usize;
-            }
-        } else {
-            while vaddr < end_vaddr {
-                let page_size = PageSize::Size4K;
-                let page = Page::new_aligned(vaddr, page_size);
-                self.map(page, paddr, flags)?;
-                vaddr += page_size as usize;
-                paddr += page_size as usize;
-            }
+        let end_vaddr = vaddr.addr() + size;
+        while vaddr.addr() < end_vaddr {
+            let page = Page::new(vaddr);
+            self.map(page, paddr, flags)?;
+            vaddr.add_offset(PAGE_SIZE);
+            paddr.add_offset(PAGE_SIZE);
         }
         Ok(())
     }
 
     fn unmap_cont(&mut self, start_vaddr: VirtAddr, size: usize) -> PagingResult {
-        assert!(is_aligned(start_vaddr));
+        assert!(is_aligned(start_vaddr.addr()));
         assert!(is_aligned(size));
-        debug!(
-            "{:#x?} unmap_cont: {:#x?}",
-            self.table_phys(),
-            start_vaddr..start_vaddr + size
-        );
         let mut vaddr = start_vaddr;
-        let end_vaddr = vaddr + size;
-        while vaddr < end_vaddr {
+        let end_vaddr = vaddr.addr() + size;
+        while vaddr.addr() < end_vaddr {
             let page_size = match self.unmap(vaddr) {
-                Ok((_, s)) => {
-                    assert!(s.is_aligned(vaddr));
-                    s as usize
+                Ok(s) => { 
+                    info!("unmap_cont: {:?}", s);
+                    PAGE_SIZE
                 }
-                Err(PagingError::NotMapped) => PageSize::Size4K as usize,
+                Err(PagingError::NotMapped) => PAGE_SIZE,
                 Err(e) => return Err(e),
             };
-            vaddr += page_size;
-            assert!(vaddr <= end_vaddr);
+            vaddr.add_offset(page_size as usize);
+            assert!(vaddr.addr() <= end_vaddr);
         }
         Ok(())
     }
 }
 
-
 impl GenericPageTable for PageTable {
     fn table_phys(&self) -> PhysAddr {
-        0
+        PhysAddr::new(0)
     }
 
     fn map(&mut self, page: Page, paddr: PhysAddr, flags: MMUFlags) -> PagingResult {
-        debug_assert!(page.size as usize == PAGE_SIZE);
-        debug_assert!(is_aligned(paddr));
-        if paddr < PMEM_SIZE {
+        debug_assert!(is_aligned(paddr.addr()));
+        if paddr.addr() < PMEM_SIZE {
             MOCK_PHYS_MEM.mmap(page.vaddr, PAGE_SIZE, paddr, flags);
             Ok(())
         } else {
@@ -211,31 +134,17 @@ impl GenericPageTable for PageTable {
         }
     }
 
-    fn unmap(&mut self, vaddr: VirtAddr) -> PagingResult<(PhysAddr, PageSize)> {
+    fn unmap(&mut self, vaddr: VirtAddr) -> PagingResult<PhysAddr> {
         self.unmap_cont(vaddr, PAGE_SIZE)?;
-        Ok((0, PageSize::Size4K))
+        Ok(PhysAddr::new(0))
     }
 
-    fn update(
-        &mut self,
-        vaddr: VirtAddr,
-        _paddr: Option<PhysAddr>,
-        flags: Option<MMUFlags>,
-    ) -> PagingResult<PageSize> {
-        debug_assert!(is_aligned(vaddr));
-        if let Some(flags) = flags {
-            MOCK_PHYS_MEM.mprotect(vaddr as _, PAGE_SIZE, flags);
-        }
-        Ok(PageSize::Size4K)
-    }
-
-    fn query(&self, vaddr: VirtAddr) -> PagingResult<(PhysAddr, MMUFlags, PageSize)> {
-        debug_assert!(is_aligned(vaddr));
-        if (PMEM_MAP_VADDR..PMEM_MAP_VADDR + PMEM_SIZE).contains(&vaddr) {
+    fn query(&self, vaddr: VirtAddr) -> PagingResult<(PhysAddr, MMUFlags)> {
+        debug_assert!(is_aligned(vaddr.addr()));
+        if (PMEM_MAP_VADDR.addr()..PMEM_MAP_VADDR.addr() + PMEM_SIZE).contains(&vaddr.addr()) {
             Ok((
-                vaddr - PMEM_MAP_VADDR,
+                PhysAddr::new(vaddr.addr() - PMEM_MAP_VADDR.addr()),
                 MMUFlags::READ | MMUFlags::WRITE,
-                PageSize::Size4K,
             ))
         } else {
             Err(PagingError::NotMapped)
@@ -246,7 +155,7 @@ impl GenericPageTable for PageTable {
         if size == 0 {
             return Ok(());
         }
-        debug_assert!(is_aligned(vaddr));
+        debug_assert!(is_aligned(vaddr.addr()));
         MOCK_PHYS_MEM.munmap(vaddr as _, size);
         Ok(())
     }
