+++ src/utils/mod.rs
@@ -1,2 +1,2 @@
 pub(crate) mod macros;
-pub(crate) mod init_once;
+pub(crate) mod once;
